+++ audio_be/src/commonTest/kotlin/ignitor/IgnitorArithmeticTest.kt
@@ -45,7 +45,7 @@ class IgnitorArithmeticTest : StringSpec({
     /** Constant-signal Ignitor for deterministic input. */
     fun const(value: Double): Ignitor = ParamIgnitor("c", value)
 
-    /** Verify every sample in [buf] is finite (no NaN, no ±Inf). */
+    /** Verify every sample in [ buf ] is finite (no NaN, no ±Inf). */
     fun AudioBuffer.allFinite(): Boolean = this.all { !it.isNaN() && !it.isInfinite() }
 
     // ═════════════════════════════════════════════════════════════════════════════
@@ -54,38 +54,38 @@ class IgnitorArithmeticTest : StringSpec({
 
     "neg flips sign" {
         val out = render(const(0.7).neg())
-        out[0].toDouble() shouldBe (-0.7 plusOrMinus 1e-6)
+        out[0] shouldBe (-0.7 plusOrMinus 1e-6)
     }
 
     "neg of negative is positive" {
         val out = render(const(-0.4).neg())
-        out[0].toDouble() shouldBe (0.4 plusOrMinus 1e-6)
+        out[0] shouldBe (0.4 plusOrMinus 1e-6)
     }
 
     "abs of negative becomes positive" {
         val out = render(const(-0.6).abs())
-        out[0].toDouble() shouldBe (0.6 plusOrMinus 1e-6)
+        out[0] shouldBe (0.6 plusOrMinus 1e-6)
     }
 
     "abs of positive stays positive" {
         val out = render(const(0.3).abs())
-        out[0].toDouble() shouldBe (0.3 plusOrMinus 1e-6)
+        out[0] shouldBe (0.3 plusOrMinus 1e-6)
     }
 
     "sq is x squared" {
         val out = render(const(0.5).sq())
-        out[0].toDouble() shouldBe (0.25 plusOrMinus 1e-6)
+        out[0] shouldBe (0.25 plusOrMinus 1e-6)
     }
 
     "sq of negative is positive (no signed-magnitude)" {
         val out = render(const(-0.5).sq())
-        out[0].toDouble() shouldBe (0.25 plusOrMinus 1e-6)
+        out[0] shouldBe (0.25 plusOrMinus 1e-6)
     }
 
     "neg, abs, sq combined chain" {
         // (-0.5).neg().abs().sq() = ((+0.5).abs()).sq() = 0.25
         val out = render(const(-0.5).neg().abs().sq())
-        out[0].toDouble() shouldBe (0.25 plusOrMinus 1e-6)
+        out[0] shouldBe (0.25 plusOrMinus 1e-6)
     }
 
     // ═════════════════════════════════════════════════════════════════════════════
@@ -94,47 +94,47 @@ class IgnitorArithmeticTest : StringSpec({
 
     "minus subtracts" {
         val out = render(const(0.7).minus(const(0.3)))
-        out[0].toDouble() shouldBe (0.4 plusOrMinus 1e-6)
+        out[0] shouldBe (0.4 plusOrMinus 1e-6)
     }
 
     "minus of itself is zero" {
         val out = render(const(0.5).minus(const(0.5)))
-        out[0].toDouble() shouldBe (0.0 plusOrMinus 1e-6)
+        out[0] shouldBe (0.0 plusOrMinus 1e-6)
     }
 
     "min picks smaller" {
         val out = render(const(0.7).min(const(0.3)))
-        out[0].toDouble() shouldBe (0.3 plusOrMinus 1e-6)
+        out[0] shouldBe (0.3 plusOrMinus 1e-6)
     }
 
     "min handles negative" {
         val out = render(const(-0.5).min(const(0.5)))
-        out[0].toDouble() shouldBe (-0.5 plusOrMinus 1e-6)
+        out[0] shouldBe (-0.5 plusOrMinus 1e-6)
     }
 
     "max picks larger" {
         val out = render(const(0.7).max(const(0.3)))
-        out[0].toDouble() shouldBe (0.7 plusOrMinus 1e-6)
+        out[0] shouldBe (0.7 plusOrMinus 1e-6)
     }
 
     "max handles negative" {
         val out = render(const(-0.5).max(const(-0.9)))
-        out[0].toDouble() shouldBe (-0.5 plusOrMinus 1e-6)
+        out[0] shouldBe (-0.5 plusOrMinus 1e-6)
     }
 
     "clamp bounds within range" {
         val out = render(const(2.0).clamp(const(-1.0), const(1.0)))
-        out[0].toDouble() shouldBe (1.0 plusOrMinus 1e-6)
+        out[0] shouldBe (1.0 plusOrMinus 1e-6)
     }
 
     "clamp passes through if in range" {
         val out = render(const(0.3).clamp(const(-1.0), const(1.0)))
-        out[0].toDouble() shouldBe (0.3 plusOrMinus 1e-6)
+        out[0] shouldBe (0.3 plusOrMinus 1e-6)
     }
 
     "clamp lower bound" {
         val out = render(const(-2.0).clamp(const(-1.0), const(1.0)))
-        out[0].toDouble() shouldBe (-1.0 plusOrMinus 1e-6)
+        out[0] shouldBe (-1.0 plusOrMinus 1e-6)
     }
 
     // ═════════════════════════════════════════════════════════════════════════════
@@ -143,23 +143,23 @@ class IgnitorArithmeticTest : StringSpec({
 
     "pow positive base" {
         val out = render(const(2.0).pow(const(3.0)))
-        out[0].toDouble() shouldBe (8.0 plusOrMinus 1e-4)
+        out[0] shouldBe (8.0 plusOrMinus 1e-4)
     }
 
     "pow negative base — signed magnitude" {
         // (-2)^3 with signed-magnitude = -(2^3) = -8 (matches real cube)
         val out = render(const(-2.0).pow(const(3.0)))
-        out[0].toDouble() shouldBe (-8.0 plusOrMinus 1e-4)
+        out[0] shouldBe (-8.0 plusOrMinus 1e-4)
     }
 
     "pow negative base with even exp differs from sq()" {
         // signed-magnitude pow(2) gives -|x|² for negative bases, NOT |x|²
         // This is the documented quirk — see numerical-safety doc.
         val out = render(const(-2.0).pow(const(2.0)))
-        out[0].toDouble() shouldBe (-4.0 plusOrMinus 1e-4)
+        out[0] shouldBe (-4.0 plusOrMinus 1e-4)
         // sq() correctly gives positive
         val sqOut = render(const(-2.0).sq())
-        sqOut[0].toDouble() shouldBe (4.0 plusOrMinus 1e-4)
+        sqOut[0] shouldBe (4.0 plusOrMinus 1e-4)
     }
 
     "pow with zero base is finite (no NaN/Inf)" {
@@ -180,12 +180,12 @@ class IgnitorArithmeticTest : StringSpec({
 
     "exp of 0 is 1" {
         val out = render(const(0.0).exp())
-        out[0].toDouble() shouldBe (1.0 plusOrMinus 1e-6)
+        out[0] shouldBe (1.0 plusOrMinus 1e-6)
     }
 
     "exp of 1 is e" {
         val out = render(const(1.0).exp())
-        out[0].toDouble() shouldBe (kexp(1.0) plusOrMinus 1e-5)
+        out[0] shouldBe (kexp(1.0) plusOrMinus 1e-5)
     }
 
     "exp of large value clamped to SAFE_MAX (no Inf)" {
@@ -197,13 +197,13 @@ class IgnitorArithmeticTest : StringSpec({
 
     "log of positive matches ln" {
         val out = render(const(2.0).log())
-        out[0].toDouble() shouldBe (ln(2.0) plusOrMinus 1e-5)
+        out[0] shouldBe (ln(2.0) plusOrMinus 1e-5)
     }
 
     "log of negative is signed-magnitude" {
         // signed-magnitude log: -ln(|x|) for x < 0
         val out = render(const(-2.0).log())
-        out[0].toDouble() shouldBe (-ln(2.0) plusOrMinus 1e-5)
+        out[0] shouldBe (-ln(2.0) plusOrMinus 1e-5)
     }
 
     "log of 0 is 0 (no -Inf)" {
@@ -214,13 +214,13 @@ class IgnitorArithmeticTest : StringSpec({
 
     "sqrt of positive matches √" {
         val out = render(const(4.0).sqrt())
-        out[0].toDouble() shouldBe (2.0 plusOrMinus 1e-5)
+        out[0] shouldBe (2.0 plusOrMinus 1e-5)
     }
 
     "sqrt of negative is signed-magnitude" {
         // sqrt(-4) = -sqrt(4) = -2 (signed-magnitude, no NaN)
         val out = render(const(-4.0).sqrt())
-        out[0].toDouble() shouldBe (-2.0 plusOrMinus 1e-5)
+        out[0] shouldBe (-2.0 plusOrMinus 1e-5)
         out.allFinite() shouldBe true
     }
 
@@ -238,13 +238,13 @@ class IgnitorArithmeticTest : StringSpec({
 
     "tanh saturates positive" {
         val out = render(const(5.0).tanh())
-        out[0].toDouble() shouldBe (ktanh(5.0) plusOrMinus 1e-5)
+        out[0] shouldBe (ktanh(5.0) plusOrMinus 1e-5)
         out[0].shouldBeLessThan(1.0001)
     }
 
     "tanh saturates negative" {
         val out = render(const(-5.0).tanh())
-        out[0].toDouble() shouldBe (ktanh(-5.0) plusOrMinus 1e-5)
+        out[0] shouldBe (ktanh(-5.0) plusOrMinus 1e-5)
         out[0].shouldBeGreaterThan(-1.0001)
     }
 
@@ -254,47 +254,47 @@ class IgnitorArithmeticTest : StringSpec({
 
     "lerp at t=0 is left" {
         val out = render(const(0.2).lerp(const(0.8), const(0.0)))
-        out[0].toDouble() shouldBe (0.2 plusOrMinus 1e-6)
+        out[0] shouldBe (0.2 plusOrMinus 1e-6)
     }
 
     "lerp at t=1 is right" {
         val out = render(const(0.2).lerp(const(0.8), const(1.0)))
-        out[0].toDouble() shouldBe (0.8 plusOrMinus 1e-6)
+        out[0] shouldBe (0.8 plusOrMinus 1e-6)
     }
 
     "lerp at t=0.5 is midpoint" {
         val out = render(const(0.2).lerp(const(0.8), const(0.5)))
-        out[0].toDouble() shouldBe (0.5 plusOrMinus 1e-6)
+        out[0] shouldBe (0.5 plusOrMinus 1e-6)
     }
 
     "range maps -1..1 to lo..hi at midpoint" {
         // input = 0 (midpoint of [-1,1]) → output = midpoint of [lo, hi]
         val out = render(const(0.0).range(const(2.0), const(10.0)))
-        out[0].toDouble() shouldBe (6.0 plusOrMinus 1e-5)
+        out[0] shouldBe (6.0 plusOrMinus 1e-5)
     }
 
     "range maps -1 to lo" {
         val out = render(const(-1.0).range(const(2.0), const(10.0)))
-        out[0].toDouble() shouldBe (2.0 plusOrMinus 1e-5)
+        out[0] shouldBe (2.0 plusOrMinus 1e-5)
     }
 
     "range maps +1 to hi" {
         val out = render(const(1.0).range(const(2.0), const(10.0)))
-        out[0].toDouble() shouldBe (10.0 plusOrMinus 1e-5)
+        out[0] shouldBe (10.0 plusOrMinus 1e-5)
     }
 
     "bipolar maps 0..1 to -1..1" {
         // 0 → -1, 1 → +1, 0.5 → 0
-        render(const(0.0).bipolar())[0].toDouble() shouldBe (-1.0 plusOrMinus 1e-6)
-        render(const(1.0).bipolar())[0].toDouble() shouldBe (1.0 plusOrMinus 1e-6)
-        render(const(0.5).bipolar())[0].toDouble() shouldBe (0.0 plusOrMinus 1e-6)
+        render(const(0.0).bipolar())[0] shouldBe (-1.0 plusOrMinus 1e-6)
+        render(const(1.0).bipolar())[0] shouldBe (1.0 plusOrMinus 1e-6)
+        render(const(0.5).bipolar())[0] shouldBe (0.0 plusOrMinus 1e-6)
     }
 
     "unipolar maps -1..1 to 0..1" {
         // -1 → 0, +1 → 1, 0 → 0.5
-        render(const(-1.0).unipolar())[0].toDouble() shouldBe (0.0 plusOrMinus 1e-6)
-        render(const(1.0).unipolar())[0].toDouble() shouldBe (1.0 plusOrMinus 1e-6)
-        render(const(0.0).unipolar())[0].toDouble() shouldBe (0.5 plusOrMinus 1e-6)
+        render(const(-1.0).unipolar())[0] shouldBe (0.0 plusOrMinus 1e-6)
+        render(const(1.0).unipolar())[0] shouldBe (1.0 plusOrMinus 1e-6)
+        render(const(0.0).unipolar())[0] shouldBe (0.5 plusOrMinus 1e-6)
     }
 
     // ═════════════════════════════════════════════════════════════════════════════
@@ -302,25 +302,25 @@ class IgnitorArithmeticTest : StringSpec({
     // ═════════════════════════════════════════════════════════════════════════════
 
     "floor rounds down" {
-        render(const(2.7).floor())[0].toDouble() shouldBe 2.0
-        render(const(-1.3).floor())[0].toDouble() shouldBe -2.0
+        render(const(2.7).floor())[0] shouldBe 2.0
+        render(const(-1.3).floor())[0] shouldBe -2.0
     }
 
     "ceil rounds up" {
-        render(const(2.3).ceil())[0].toDouble() shouldBe 3.0
-        render(const(-1.7).ceil())[0].toDouble() shouldBe -1.0
+        render(const(2.3).ceil())[0] shouldBe 3.0
+        render(const(-1.7).ceil())[0] shouldBe -1.0
     }
 
     "round to nearest integer" {
-        render(const(2.4).round())[0].toDouble() shouldBe 2.0
-        render(const(2.6).round())[0].toDouble() shouldBe 3.0
-        render(const(-1.4).round())[0].toDouble() shouldBe -1.0
+        render(const(2.4).round())[0] shouldBe 2.0
+        render(const(2.6).round())[0] shouldBe 3.0
+        render(const(-1.4).round())[0] shouldBe -1.0
     }
 
     "frac is x minus floor(x)" {
-        render(const(2.7).frac())[0].toDouble() shouldBe (0.7 plusOrMinus 1e-6)
-        render(const(-1.3).frac())[0].toDouble() shouldBe (0.7 plusOrMinus 1e-6)  // -1.3 - (-2) = 0.7
-        render(const(0.0).frac())[0].toDouble() shouldBe 0.0
+        render(const(2.7).frac())[0] shouldBe (0.7 plusOrMinus 1e-6)
+        render(const(-1.3).frac())[0] shouldBe (0.7 plusOrMinus 1e-6)  // -1.3 - (-2) = 0.7
+        render(const(0.0).frac())[0] shouldBe 0.0
     }
 
     // ═════════════════════════════════════════════════════════════════════════════
@@ -329,9 +329,9 @@ class IgnitorArithmeticTest : StringSpec({
 
     "mod follows Kotlin rem semantics — sign follows dividend" {
         // 5.5 % 2.0 = 1.5
-        render(const(5.5).mod(const(2.0)))[0].toDouble() shouldBe (1.5 plusOrMinus 1e-5)
+        render(const(5.5).mod(const(2.0)))[0] shouldBe (1.5 plusOrMinus 1e-5)
         // -5.5 % 2.0 = -1.5 (rem, not floored mod)
-        render(const(-5.5).mod(const(2.0)))[0].toDouble() shouldBe (-1.5 plusOrMinus 1e-5)
+        render(const(-5.5).mod(const(2.0)))[0] shouldBe (-1.5 plusOrMinus 1e-5)
     }
 
     "mod by zero is finite (epsilon-substituted divisor)" {
@@ -340,8 +340,8 @@ class IgnitorArithmeticTest : StringSpec({
     }
 
     "recip is 1/x" {
-        render(const(2.0).recip())[0].toDouble() shouldBe (0.5 plusOrMinus 1e-5)
-        render(const(-4.0).recip())[0].toDouble() shouldBe (-0.25 plusOrMinus 1e-5)
+        render(const(2.0).recip())[0] shouldBe (0.5 plusOrMinus 1e-5)
+        render(const(-4.0).recip())[0] shouldBe (-0.25 plusOrMinus 1e-5)
     }
 
     "recip of zero is finite (clamped to SAFE_MAX)" {
@@ -376,7 +376,7 @@ class IgnitorArithmeticTest : StringSpec({
 
     "times multiplies normally" {
         val out = render(const(0.5) * const(0.4))
-        out[0].toDouble() shouldBe (0.2 plusOrMinus 1e-6)
+        out[0] shouldBe (0.2 plusOrMinus 1e-6)
     }
 
     "times of huge values clamped to SAFE_MAX (no Inf)" {
@@ -391,17 +391,17 @@ class IgnitorArithmeticTest : StringSpec({
 
     "select picks whenTrue when cond > 0" {
         val out = render(const(0.5).select(const(7.0), const(3.0)))
-        out[0].toDouble() shouldBe (7.0 plusOrMinus 1e-6)
+        out[0] shouldBe (7.0 plusOrMinus 1e-6)
     }
 
     "select picks whenFalse when cond <= 0" {
         val out = render(const(-0.5).select(const(7.0), const(3.0)))
-        out[0].toDouble() shouldBe (3.0 plusOrMinus 1e-6)
+        out[0] shouldBe (3.0 plusOrMinus 1e-6)
     }
 
     "select cond=0 picks whenFalse (strict > 0 semantics)" {
         val out = render(const(0.0).select(const(7.0), const(3.0)))
-        out[0].toDouble() shouldBe (3.0 plusOrMinus 1e-6)
+        out[0] shouldBe (3.0 plusOrMinus 1e-6)
     }
 
     // ═════════════════════════════════════════════════════════════════════════════
